スキル中身作れなかった

+++ sources/EnemySpawner.cpp
@@ -27,7 +27,7 @@ EnemyBase* EnemySpawner::spawnEnemy(GameObject* pParent, EnemyType type, bool is
 	auto enemy = EnemyFactoty::createEnemy(pParent, type, isBoss);
 	if (!enemy)
 		assert(false);
-	XMMATRIX rotmat = XMMatrixRotationY(XMConvertToRadians(rand() % ANGLE360));
+	XMMATRIX rotmat = XMMatrixRotationY(XMConvertToRadians((float)(rand() % ANGLE360)));
 	XMVECTOR vpos = XMVector3TransformCoord(XMVectorSet(0,0, SPAWN_DISTANCE,0), rotmat);
 	enemy->SetPosition(vpos);
 	return enemy;
